diagnostic for missing files in import statements. dedicated import spec
