+++ src/Scope.ts
@@ -1,10 +1,10 @@
 import { EventEmitter } from 'eventemitter3';
 import { CompletionItem, CompletionItemKind, Location, Position, Range } from 'vscode-languageserver';
 import chalk from 'chalk';
-import { DiagnosticMessages } from './DiagnosticMessages';
+import { DiagnosticMessages, DiagnosticInfo } from './DiagnosticMessages';
 import { BrsFile } from './files/BrsFile';
 import { XmlFile } from './files/XmlFile';
-import { CallableContainer, BsDiagnostic } from './interfaces';
+import { CallableContainer, BsDiagnostic, FileReference } from './interfaces';
 import { Program } from './Program';
 import { BsClassValidator } from './validators/ClassValidator';
 import { NamespaceStatement, ParseMode, Statement, NewExpression, FunctionStatement } from './parser';
@@ -358,6 +358,9 @@ export class Scope {
         //find all duplicate function declarations
         this.diagnosticFindDuplicateFunctionDeclarations(callableContainerMap);
 
+        //detect missing and incorrect-case script imports
+        this.diagnosticValidateScriptImportPaths();
+
         //enforce a series of checks on the bodies of class methods
         this.validateClasses();
 
@@ -657,6 +660,55 @@ export class Scope {
                         file: callable.file
                     });
                 }
+            }
+        }
+    }
+
+    /**
+     * Get the list of all script imports for this scope
+     */
+    private getScriptImports() {
+        let result = [] as FileReference[];
+        let files = this.getFiles();
+        for (let file of files) {
+            if (file instanceof BrsFile) {
+                result.push(...file.ownScriptImports);
+            } else if (file instanceof XmlFile) {
+                result.push(...file.scriptTagImports);
+            }
+        }
+        return result;
+    }
+
+    /**
+     * Verify that all of the scripts ipmorted by
+     */
+    private diagnosticValidateScriptImportPaths() {
+        let scriptImports = this.getScriptImports();
+        //verify every script import
+        for (let scriptImport of scriptImports) {
+            let referencedFile = this.getFileByRelativePath(scriptImport.pkgPath);
+            //if we can't find the file
+            if (!referencedFile) {
+                let dInfo: DiagnosticInfo;
+                if (scriptImport.text.trim().length === 0) {
+                    dInfo = DiagnosticMessages.scriptSrcCannotBeEmpty();
+                } else {
+                    dInfo = DiagnosticMessages.referencedFileDoesNotExist();
+                }
+
+                this.diagnostics.push({
+                    ...dInfo,
+                    range: scriptImport.filePathRange,
+                    file: scriptImport.sourceFile
+                });
+                //if the character casing of the script import path does not match that of the actual path
+            } else if (scriptImport.pkgPath !== referencedFile.pkgPath) {
+                this.diagnostics.push({
+                    ...DiagnosticMessages.scriptImportCaseMismatch(referencedFile.pkgPath),
+                    range: scriptImport.filePathRange,
+                    file: scriptImport.sourceFile
+                });
             }
         }
     }
