+++ src/files/BrsFile.ts
@@ -209,7 +209,7 @@ export class BrsFile {
                     filePathRange: result.filePathToken.range,
                     pkgPath: util.getPkgPathFromTarget(this.pkgPath, result.filePath),
                     sourceFile: this,
-                    text: ''
+                    text: result.filePathToken?.text
                 });
             }
